+++ clients/web3-wallet/src/binance_sdk_web3_wallet/rest_api/models/get_transaction_status_response_data.py
@@ -15,7 +15,6 @@ import pprint
 import re  # noqa: F401
 import json
 
-
 from pydantic import BaseModel, ConfigDict, Field, StrictStr
 from typing import Any, ClassVar, Dict, List, Optional
 from binance_sdk_web3_wallet.rest_api.models.get_transaction_status_response_data_from_token_details_inner import (